requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.2.1
python-dotenv